Refactor in order to add tests for assigning jobs
