+++ application.rb
@@ -2,11 +2,12 @@ require 'sinatra'
 require 'active_support/time'
 
 require './lib/calculate_week'
+require './lib/assign_jobs_per_week'
 
 before do
   current_week_commencing = Date.today.beginning_of_week(:sunday).strftime('%d/%m/%Y')
   @week = calculate_week(current_week_commencing)
-  assign_jobs_per_week
+  assign_jobs_per_week # Returns `@person_week_n` variables.
 end
 
 get '/' do
@@ -27,37 +28,3 @@ get '/jobs.json' do
   end
 end
 
-def jobs
-  [
-    'hoover lounge',
-    'dust/wipe lounge',
-    'empty lounge bin',
-    'hoover stairs and landing',
-    'hoover and mop porch and downstairs bathroom',
-    'clean toilet and sink downstairs',
-    'hoover and mop kitchen floor',
-    'garden',
-    'wipe dining table and sides',
-    'clean hobs and sink',
-    'clean microwave',
-    'empty bins',
-    'clean toilet and sink upstairs',
-    'clean bath and shower screen',
-    'clean bathroom surfaces and mirror',
-    'hoover upstairs landing and hoover and mop bathroom floor',
-  ]
-end
-
-def assign_jobs_per_week
-  @jack_week_a = [jobs[0], jobs[3], jobs[4], jobs[6], jobs[15]]
-  @jack_week_b = [jobs[7], jobs[8], jobs[9], jobs[10], jobs[11]]
-  @jack_week_c = [jobs[1], jobs[3], jobs[5], jobs[12], jobs[13], jobs[14]]
-
-  @phil_week_a = [jobs[1], jobs[2], jobs[5], jobs[12], jobs[13], jobs[14]]
-  @phil_week_b = [jobs[0], jobs[3], jobs[4], jobs[6], jobs[15]]
-  @phil_week_c = [jobs[7], jobs[8], jobs[9], jobs[10], jobs[11]]
-
-  @isabell_week_a = [jobs[7], jobs[8], jobs[9], jobs[10], jobs[11]]
-  @isabell_week_b = [jobs[1], jobs[2], jobs[5], jobs[12], jobs[13], jobs[14]]
-  @isabell_week_c = [jobs[0], jobs[3], jobs[4], jobs[6], jobs[15]]
-end
